Added %colheightleft size spec, and fleshed out size spec docs and implementation
The documentation was left unfinished for the size specifications.
Furthermore, the custom size spec handling was only partially
implemented. This commit finishes the documentation and finalizes
the implementation. It also adds the %colheightleft size spec. The
speck only pays attention to packed widgets, so it could make
mistakes, but it properly handles packing on the bottom or top
of the container.