Improved alignment inheritance and vastly improvied font hash handling
In particular, I finally bit the bullet and figured out how to merge
font hashes. In the process, I also vastly expanded the size and
height specs that can be used for fonts in renderables! I also
deprecated the font_factor field for paragraphs, since that can
now be achieved with the new font size spec handling. For example,
to achieve the old font_factor => 1.2, use font => { size => '1.2x' }